refactor: simplify domain naming
