+++ hals/src/http/cors.ts
