+++ hals/src/shared/number-range.ts
