+++ hals/src/shared/page.ts
