+++ hals/src/shared/param-map.ts
