+++ hals/src/shared/result.ts
